Replace refund controller after rebase
