+++ src/Controller/Api/RefundController.php
@@ -2,57 +2,27 @@
 
 namespace Kiener\MolliePayments\Controller\Api;
 
-use Exception;
-use Kiener\MolliePayments\Factory\MollieApiFactory;
-use Kiener\MolliePayments\Service\CustomFieldService;
 use Kiener\MolliePayments\Service\OrderService;
+use Kiener\MolliePayments\Service\RefundService;
 use Kiener\MolliePayments\Service\SettingsService;
-use Kiener\MolliePayments\Setting\MollieSettingStruct;
-use Mollie\Api\Exceptions\ApiException;
-use Mollie\Api\Exceptions\IncompatiblePlatform;
-use Mollie\Api\MollieApiClient;
-use Shopware\Core\Checkout\Order\Aggregate\OrderLineItem\OrderLineItemEntity;
+use Psr\Log\LoggerInterface;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Checkout\Order\OrderEntity;
+use Shopware\Core\Checkout\Payment\Exception\InvalidOrderException;
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
+use Shopware\Core\Framework\ShopwareHttpException;
+use Shopware\Core\Framework\Uuid\Exception\InvalidUuidException;
+use Shopware\Core\Framework\Uuid\Uuid;
+use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 
 class RefundController extends AbstractController
 {
-    public const CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY = 'refundedQuantity';
-    public const CUSTOM_FIELDS_KEY_CREATE_CREDIT_ITEM = 'createCredit';
-
-    private const CUSTOM_FIELDS_KEY_ORDER_ID = 'order_id';
-    private const CUSTOM_FIELDS_KEY_ORDER_LINE_ID = 'order_line_id';
-    private const CUSTOM_FIELDS_KEY_REFUND_ID = 'refund_id';
-    private const CUSTOM_FIELDS_KEY_REFUNDS = 'refunds';
-    private const CUSTOM_FIELDS_KEY_QUANTITY = 'quantity';
-
-    private const REFUND_DATA_KEY_ID = 'id';
-    private const REFUND_DATA_KEY_LINES = 'lines';
-    private const REFUND_DATA_KEY_QUANTITY = self::CUSTOM_FIELDS_KEY_QUANTITY;
-    private const REFUND_DATA_KEY_TEST_MODE = 'testmode';
-
-    private const REQUEST_KEY_ORDER_LINE_ITEM_ID = 'itemId';
-    private const REQUEST_KEY_ORDER_LINE_QUANTITY = self::CUSTOM_FIELDS_KEY_QUANTITY;
-    private const REQUEST_KEY_ORDER_LINE_ITEM_VERSION_ID = 'versionId';
-
-    private const RESPONSE_KEY_AMOUNT = 'amount';
-    private const RESPONSE_KEY_ITEMS = 'items';
-    private const RESPONSE_KEY_SUCCESS = 'success';
-
-    /** @var MollieApiFactory */
-    private $apiFactory;
-
-    /** @var EntityRepositoryInterface */
-    private $orderLineItemRepository;
+    /** @var LoggerInterface */
+    private $logger;
 
     /** @var OrderService */
     private $orderService;
@@ -63,326 +33,259 @@ class RefundController extends AbstractController
     /** @var SettingsService */
     private $settingsService;
 
+    /** @var RefundService */
+    private $refundService;
+
     /**
      * Creates a new instance of the onboarding controller.
      *
-     * @param MollieApiFactory $apiFactory
-     * @param EntityRepositoryInterface $orderLineItemRepository
+     * @param LoggerInterface $logger
      * @param OrderService $orderService
      * @param OrderTransactionStateHandler $orderTransactionStateHandler
      * @param SettingsService $settingsService
+     * @param RefundService $refundService
      */
     public function __construct(
-        MollieApiFactory $apiFactory,
-        EntityRepositoryInterface $orderLineItemRepository,
+        LoggerInterface $logger,
         OrderService $orderService,
         OrderTransactionStateHandler $orderTransactionStateHandler,
-        SettingsService $settingsService
+        SettingsService $settingsService,
+        RefundService $refundService
     )
     {
-        $this->apiFactory = $apiFactory;
-        $this->orderLineItemRepository = $orderLineItemRepository;
+        $this->logger = $logger;
         $this->orderService = $orderService;
         $this->orderTransactionStateHandler = $orderTransactionStateHandler;
         $this->settingsService = $settingsService;
+        $this->refundService = $refundService;
+    }
+
+    /**
+     * @RouteScope(scopes={"api"})
+     * @Route("/api/_action/mollie/refund",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund", methods={"POST"})
+     *
+     * @param RequestDataBag $data
+     * @param Context $context
+     * @return JsonResponse
+     */
+    public function refund(RequestDataBag $data, Context $context): JsonResponse
+    {
+        return $this->refundResponse($data->getAlnum('orderId'), $data->get('amount', 0.0), $context);
     }
 
     /**
      * @RouteScope(scopes={"api"})
      * @Route("/api/v{version}/_action/mollie/refund",
-     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund", methods={"POST"})
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.legacy", methods={"POST"})
      *
-     * @param Request $request
-     *
+     * @param RequestDataBag $data
+     * @param Context $context
      * @return JsonResponse
-     * @throws ApiException
-     * @throws IncompatiblePlatform
      */
-    public function refund(Request $request): JsonResponse
+    public function refundLegacy(RequestDataBag $data, Context $context): JsonResponse
     {
-        return $this->getRefundResponse($request);
-    }
-
-    /**
-     * refund action for Shopware versions >=6.4
-     *
-     * @RouteScope(scopes={"api"})
-     * @Route("/api/mollie/refund",
-     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund-64", methods={"POST"})
-     *
-     * @param Request $request
-     *
-     * @return JsonResponse
-     * @throws ApiException
-     * @throws IncompatiblePlatform
-     */
-    public function refund64(Request $request): JsonResponse
-    {
-        return $this->getRefundResponse($request);
-    }
-
-    private function getRefundResponse(Request $request): JsonResponse
-    {
-        /** @var MollieApiClient|null $apiClient */
-        $apiClient = null;
-
-        /** @var array|null $customFields */
-        $customFields = null;
-
-        /** @var string|null $orderId */
-        $orderId = null;
-
-        /** @var string|null $orderLineId */
-        $orderLineId = null;
-
-        /** @var OrderLineItemEntity $orderLineItem */
-        $orderLineItem = null;
-
-        /** @var bool $success */
-        $success = false;
-
-        /** @var int $quantity */
-        $quantity = 0;
-
-        if (
-            (string)$request->get(self::REQUEST_KEY_ORDER_LINE_ITEM_ID) !== ''
-            && (string)$request->get(self::REQUEST_KEY_ORDER_LINE_ITEM_VERSION_ID) !== ''
-        ) {
-            $orderLineItem = $this->getOrderLineItemById(
-                $request->get(self::REQUEST_KEY_ORDER_LINE_ITEM_ID),
-                $request->get(self::REQUEST_KEY_ORDER_LINE_ITEM_VERSION_ID)
-            );
-        }
-
-        if ((int)$request->get(self::REQUEST_KEY_ORDER_LINE_QUANTITY) > 0) {
-            $quantity = (int)$request->get(self::REQUEST_KEY_ORDER_LINE_QUANTITY);
-        }
-
-        if (
-            $orderLineItem !== null
-            && !empty($orderLineItem->getCustomFields())
-        ) {
-            $customFields = $orderLineItem->getCustomFields();
-        }
-
-        if (
-            $orderLineItem !== null
-            && !empty($customFields)
-            && isset($customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_ORDER_LINE_ID])
-        ) {
-            $orderLineId = $customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_ORDER_LINE_ID];
-        }
-
-        if (
-            $orderLineItem !== null
-            && $orderLineItem->getOrder() !== null
-            && !empty($orderLineItem->getOrder()->getCustomFields())
-            && isset($orderLineItem->getOrder()->getCustomFields()[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_ORDER_ID])
-        ) {
-            $orderId = $orderLineItem->getOrder()->getCustomFields()[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_ORDER_ID];
-        }
-
-        if ($orderLineItem->getOrder() !== null) {
-            $transactions = $orderLineItem->getOrder()->getTransactions();
-
-            if ($transactions !== null && $transactions->count()) {
-                foreach ($transactions as $transaction) {
-                    try {
-                        $this->orderTransactionStateHandler->refundPartially(
-                            $transaction->getId(),
-                            Context::createDefaultContext()
-                        );
-                    } catch (Exception $e) {
-                        // @todo Maybe handle this exception in debug mode?
-                    }
-                }
-            }
-        }
-
-        if (
-            (string)$orderId !== ''
-            && (string)$orderLineId !== ''
-            && $quantity > 0
-        ) {
-            $apiClient = $this->apiFactory->createClient(
-                $orderLineItem->getOrder()->getSalesChannelId()
-            );
-        }
-
-        if ($apiClient !== null) {
-            /** @var MollieSettingStruct $settings */
-            $settings = $this->settingsService->getSettings(
-                $orderLineItem->getOrder()->getSalesChannelId()
-            );
-
-            /** @var array $orderParameters */
-            $orderParameters = [];
-
-            if ($settings->isTestMode() && $apiClient->usesOAuth()) {
-                $orderParameters = [
-                    self::REFUND_DATA_KEY_TEST_MODE => true
-                ];
-            }
-
-            try {
-                $order = $apiClient->orders->get($orderId, $orderParameters);
-            } catch (ApiException $e) {
-                //
-            }
-
-            if (isset($order, $order->id)) {
-                $refundData = [
-                    self::REFUND_DATA_KEY_LINES => [
-                        [
-                            self::REFUND_DATA_KEY_ID => $orderLineId,
-                            self::REFUND_DATA_KEY_QUANTITY => $quantity,
-                        ],
-                    ],
-                ];
-
-                if ($settings->isTestMode() && $apiClient->usesOAuth()) {
-                    $refundData[self::REFUND_DATA_KEY_TEST_MODE] = true;
-                }
-
-                try {
-                    $refund = $apiClient->orderRefunds->createFor($order, $refundData);
-                } catch (ApiException $e) {
-                    //
-                }
-
-                if (isset($refund, $refund->id)) {
-                    $success = true;
-
-                    if (!isset($customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_REFUNDS])) {
-                        $customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_REFUNDS] = [];
-                    }
-
-                    if (!is_array($customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_REFUNDS])) {
-                        $customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_REFUNDS] = [];
-                    }
-
-                    $customFields[CustomFieldService::CUSTOM_FIELDS_KEY_MOLLIE_PAYMENTS][self::CUSTOM_FIELDS_KEY_REFUNDS][] = [
-                        self::CUSTOM_FIELDS_KEY_REFUND_ID => $refund->id,
-                        self::CUSTOM_FIELDS_KEY_QUANTITY => $quantity,
-                    ];
-
-                    if (isset($customFields[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY])) {
-                        $customFields[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY] += $quantity;
-                    } else {
-                        $customFields[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY] = $quantity;
-                    }
-                }
-            }
-
-            // Update the custom fields of the order line item
-            $this->orderLineItemRepository->update([
-                [
-                    self::REFUND_DATA_KEY_ID => $orderLineItem->getId(),
-                    CustomFieldService::CUSTOM_FIELDS_KEY => $customFields,
-                ]
-            ], Context::createDefaultContext());
-        }
-
-        return new JsonResponse([
-            self::RESPONSE_KEY_SUCCESS => $success,
-        ]);
+        return $this->refundResponse($data->getAlnum('orderId'), $data->get('amount', 0.0), $context);
     }
 
     /**
      * @RouteScope(scopes={"api"})
-     * @Route("/api/v{version}/_action/mollie/refund/total",
-     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.total", methods={"POST"})
+     * @Route("/api/_action/mollie/refund/cancel",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.cancel", methods={"POST"})
      *
-     * @param Request $request
-     *
+     * @param RequestDataBag $data
+     * @param Context $context
      * @return JsonResponse
      */
-    public function total(Request $request): JsonResponse
+    public function cancel(RequestDataBag $data, Context $context): JsonResponse
     {
-        /** @var string|null $orderId */
-        $orderId = $request->get('orderId');
+        return $this->cancelResponse($data->getAlnum('orderId'), $data->get('refundId'), $context);
+    }
 
-        return $this->getTotalResponse($orderId);
+    /**
+     * @RouteScope(scopes={"api"})
+     * @Route("/api/v{version}/_action/mollie/refund/cancel",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.cancel.legacy", methods={"POST"})
+     *
+     * @param RequestDataBag $data
+     * @param Context $context
+     * @return JsonResponse
+     */
+    public function cancelLegacy(RequestDataBag $data, Context $context): JsonResponse
+    {
+        return $this->cancelResponse($data->getAlnum('orderId'), $data->get('refundId'), $context);
+    }
+
+    /**
+     * @RouteScope(scopes={"api"})
+     * @Route("/api/_action/mollie/refund/list",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.list", methods={"POST"})
+     *
+     * @param RequestDataBag $data
+     * @param Context $context
+     * @return JsonResponse
+     */
+    public function list(RequestDataBag $data, Context $context): JsonResponse
+    {
+        return $this->listResponse($data->getAlnum('orderId'), $context);
+    }
+
+    /**
+     * @RouteScope(scopes={"api"})
+     * @Route("/api/v{version}/_action/mollie/refund/list",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.list.legacy", methods={"POST"})
+     *
+     * @param RequestDataBag $data
+     * @param Context $context
+     * @return JsonResponse
+     */
+    public function listLegacy(RequestDataBag $data, Context $context): JsonResponse
+    {
+        return $this->listResponse($data->getAlnum('orderId'), $context);
     }
 
     /**
      * @RouteScope(scopes={"api"})
      * @Route("/api/_action/mollie/refund/total",
-     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.total-64", methods={"POST"})
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.total", methods={"POST"})
      *
-     * @param Request $request
-     *
+     * @param RequestDataBag $data
+     * @param Context $context
      * @return JsonResponse
      */
-    public function total64(Request $request): JsonResponse
+    public function total(RequestDataBag $data, Context $context): JsonResponse
     {
-        /** @var string|null $orderId */
-        $orderId = $request->get('orderId');
-
-        return $this->getTotalResponse($orderId);
+        return $this->totalResponse($data->getAlnum('orderId'), $context);
     }
 
-    private function getTotalResponse(?string $orderId): JsonResponse
+    /**
+     * @RouteScope(scopes={"api"})
+     * @Route("/api/v{version}/_action/mollie/refund/total",
+     *         defaults={"auth_enabled"=true}, name="api.action.mollie.refund.total.legacy", methods={"POST"})
+     *
+     * @param RequestDataBag $data
+     * @param Context $context
+     * @return JsonResponse
+     */
+    public function totalLegacy(RequestDataBag $data, Context $context): JsonResponse
     {
-        /** @var float $amount */
-        $amount = 0.0;
+        return $this->totalResponse($data->getAlnum('orderId'), $context);
+    }
 
-        /** @var int $items */
-        $items = 0;
+    /**
+     * @param string $orderId
+     * @param float $amount
+     * @param Context $context
+     * @return JsonResponse
+     */
+    private function refundResponse(string $orderId, float $amount, Context $context): JsonResponse
+    {
+        try {
+            $order = $this->getValidOrder($orderId, $context);
 
-        /** @var OrderEntity $order */
-        $order = null;
-
-        if (!empty($orderId)) {
-            $order = $this->orderService->getOrder($orderId, Context::createDefaultContext());
+            $success = $this->refundService->refund($order, $amount);
+        } catch (ShopwareHttpException $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], $e->getStatusCode());
+        } catch (\Throwable $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], 500);
         }
 
-        if ($order !== null) {
-            foreach ($order->getLineItems() as $lineItem) {
-                if (
-                    !empty($lineItem->getCustomFields())
-                    && isset($lineItem->getCustomFields()[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY])
-                ) {
-                    $amount += ($lineItem->getUnitPrice() * (int)$lineItem->getCustomFields()[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY]);
-                    $items += (int)$lineItem->getCustomFields()[self::CUSTOM_FIELDS_KEY_REFUNDED_QUANTITY];
-                }
-            }
-        }
-
-        return new JsonResponse([
-            self::RESPONSE_KEY_AMOUNT => $amount,
-            self::RESPONSE_KEY_ITEMS => $items,
+        return $this->json([
+            'success' => $success
         ]);
     }
 
     /**
-     * Returns an order line item by id.
-     *
-     * @param              $lineItemId
-     * @param null $versionId
-     * @param Context|null $context
-     *
-     * @return OrderLineItemEntity|null
+     * @param string $orderId
+     * @param string|null $refundId
+     * @param Context $context
+     * @return JsonResponse
      */
-    public function getOrderLineItemById(
-        $lineItemId,
-        $versionId = null,
-        Context $context = null
-    ): ?OrderLineItemEntity
+    private function cancelResponse(string $orderId, ?string $refundId, Context $context): JsonResponse
     {
-        $orderLineCriteria = new Criteria([$lineItemId]);
+        try {
+            $order = $this->getValidOrder($orderId, $context);
 
-        if ($versionId !== null) {
-            $orderLineCriteria->addFilter(new EqualsFilter('versionId', $versionId));
+            $success = $this->refundService->cancel($order, $refundId);
+        } catch (ShopwareHttpException $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], $e->getStatusCode());
+        } catch (\Throwable $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], 500);
         }
 
-        $orderLineCriteria->addAssociation('order');
-        $orderLineCriteria->addAssociation('order.salesChannel');
-        $orderLineCriteria->addAssociation('order.transactions');
+        return $this->json([
+            'success' => $success
+        ]);
+    }
 
-        return $this->orderLineItemRepository->search(
-            $orderLineCriteria,
-            $context ?? Context::createDefaultContext()
-        )->get($lineItemId);
+    /**
+     * @param string $orderId
+     * @param Context $context
+     * @return JsonResponse
+     */
+    private function listResponse(string $orderId, Context $context): JsonResponse
+    {
+        try {
+            $order = $this->getValidOrder($orderId, $context);
+
+            $refunds = $this->refundService->getRefunds($order);
+        } catch (ShopwareHttpException $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], $e->getStatusCode());
+        } catch (\Throwable $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], 500);
+        }
+
+        return $this->json($refunds ?? []);
+    }
+
+    /**
+     * @param string $orderId
+     * @param Context $context
+     * @return JsonResponse
+     */
+    private function totalResponse(string $orderId, Context $context): JsonResponse
+    {
+        try {
+            $order = $this->getValidOrder($orderId, $context);
+
+            $this->refundService->getRefunds($order);
+
+            $remaining = $this->refundService->getRemainingAmount($order);
+            $refunded = $this->refundService->getRefundedAmount($order);
+        } catch (ShopwareHttpException $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], $e->getStatusCode());
+        } catch (\Throwable $e) {
+            $this->logger->error($e->getMessage());
+            return $this->json(['message' => $e->getMessage()], 404);
+        }
+
+        return $this->json(compact('remaining', 'refunded'));
+    }
+
+    /**
+     * @param string $orderId
+     * @param Context $context
+     * @return OrderEntity
+     */
+    private function getValidOrder(string $orderId, Context $context): OrderEntity
+    {
+        if (!Uuid::isValid($orderId)) {
+            throw new InvalidUuidException($orderId);
+        }
+
+        $order = $this->orderService->getOrder($orderId, $context);
+
+        if (!($order instanceof OrderEntity)) {
+            throw new InvalidOrderException($orderId);
+        }
+
+        return $order;
     }
 }
